xlsx_extractor/requirements.txt
pandas
openpyxl
flask
flask_cors
requests
mysql-connector-python
python-dotenv
